+++ package.json
@@ -18,6 +18,7 @@
     "apollo-boost": "^0.1.16",
     "apollo-link-context": "^1.0.9",
     "apollo-link-debounce": "^2.1.0",
+    "apollo-link-watched-mutation": "^0.2.4",
     "draft-js": "^0.10.5",
     "emotion": "^9.2.12",
     "escape-html": "^1.0.3",
