+++ src/components/CreateTask/index.js
@@ -1,18 +1,26 @@
 import React from 'react';
 import {useMutation} from 'react-apollo-hooks';
+import styled from '@emotion/styled/macro';
 
 import TaskInput from '../TaskInput';
+import {TaskContainer} from '../TasksList/task';
 
 import {ADD_ITEM} from '../../utils/mutations';
 
-const CreateTask = () => {
-	const createTask = useMutation(ADD_ITEM, {
-		update: (proxy, result) => {
-			// update tasks query
-		},
-	});
+const TaskInputContainer = styled('div')`
+	& + ${TaskContainer} {
+		margin-top: 3rem;
+	}
+`;
 
-	return <TaskInput onSubmitTask={task => createTask({variables: task})} />;
+const CreateTask = () => {
+	const createTask = useMutation(ADD_ITEM);
+
+	return (
+		<TaskInputContainer>
+			<TaskInput onSubmitTask={task => createTask({variables: task})} />
+		</TaskInputContainer>
+	);
 };
 
 export default CreateTask;
