+++ src/components/TaskInput/index.js
@@ -5,6 +5,28 @@ import useOnClickOutside from 'use-onclickoutside';
 
 import TaskTypeDropdown from '../TaskTypeDropdown';
 
+import TaskIconUrl, {
+	ReactComponent as TaskIcon,
+} from '../../utils/icons/taskicon.svg';
+import TaskIconValidatedUrl from '../../utils/icons/taskicon-user-validated.svg';
+import TaskCustomerIconUrl, {
+	ReactComponent as TaskCustomerIcon,
+} from '../../utils/icons/taskicon-customer.svg';
+import TaskCustomerIconValidatedUrl from '../../utils/icons/taskicon-customer-validated.svg';
+import CustomerReminderIconUrl, {
+	ReactComponent as CustomerReminderIcon,
+} from '../../utils/icons/customer-reminder.svg';
+import UserReminderIconUrl, {
+	ReactComponent as UserReminderIcon,
+} from '../../utils/icons/user-reminder.svg';
+import ValidationIconUrl, {
+	ReactComponent as ValidationIcon,
+} from '../../utils/icons/validation.svg';
+import ContentAcquisitionIconUrl, {
+	ReactComponent as ContentAcquisitionIcon,
+} from '../../utils/icons/content-acquisition.svg';
+import {ITEM_TYPES} from '../../utils/constants.js';
+
 const Container = styled('div')`
 	font-size: 14px;
 `;
@@ -21,7 +43,7 @@ const Input = styled('input')`
 	border: none;
 	border-radius: 20px;
 	padding: 0.5rem 1.2rem 0.5rem 4rem;
-	margin-left: -2.1rem;
+	margin-left: -2.2rem;
 	color: #5020ee;
 	font-size: 18px;
 	border: 1px solid transparent;
@@ -47,94 +69,25 @@ const Icon = styled('div')`
 	display: flex;
 	justify-content: center;
 	align-items: center;
-	background-color: #5020ee;
+	background-color: ${props => (props.active ? 'transparent' : '#5020ee')};
 	color: #fff;
 	border: 2px solid transparent;
 	border-radius: 50%;
-	min-width: 1.3rem;
-	min-height: 1.3rem;
-	width: 1.3rem;
-	height: 1.3rem;
+	width: 26px;
+	height: 26px;
 	z-index: 1;
 	transition: all 400ms ease;
 	cursor: pointer;
 
 	&:hover {
-		border: 2px solid #5020ee;
+		border: 2px solid ${props => (props.active ? 'transparent' : '#5020ee')};
 		color: #5020ee;
 		background-color: #fff;
 		transition: all 400ms ease;
 	}
 `;
 
-const types = [
-	{
-		icon: '🔵',
-		type: 'DEFAULT',
-		name: 'Tâche par défaut',
-		description: 'Une tâche dont vous êtes responsable',
-	},
-	{
-		icon: '🔴',
-		type: 'CUSTOMER',
-		name: 'Tâche attribuée au client',
-		description: 'Une tâche à réaliser par votre client',
-	},
-	{
-		icon: '⏰',
-		type: 'REMINDER',
-		name: 'Relance client générique',
-		description: 'Programmer des relances client',
-	},
-	{
-		icon: '👍',
-		type: 'VALIDATION',
-		name: 'Validation client',
-		description: 'Demander à votre client une validation',
-	},
-	{
-		icon: '🔔',
-		type: 'PERSONAL_REMINDER',
-		name: 'Rappel personnel',
-		description: 'Programmer un rappel (visible seulement par vous)',
-	},
-	{
-		icon: '📝',
-		type: 'MEETING_NOTES',
-		name: 'Réunion client',
-		description: 'Assembler et partager les notes de réunion',
-	},
-	{
-		icon: '📁',
-		type: 'CONTENT_ACQUISITION',
-		name: 'Récupération contenu',
-		description: 'Lister et récupérer les contenus nécessaires',
-	},
-	{
-		icon: '🌳',
-		type: 'SUBTASKS',
-		name: 'Tâche et sous-tâches',
-		description: "Lister les sous-tâches d'une tâche parente",
-	},
-	{
-		icon: '💰',
-		type: 'PAYMENT',
-		name: 'Paiement par le client',
-		description: 'Demander et relancer pour un paiement',
-	},
-	{
-		icon: '📆',
-		type: 'SCHEDULE_MEETING',
-		name: 'Programmation de RDV client',
-		description: 'Programmer automatiquement une réunion',
-	},
-	{
-		icon: '⭕',
-		type: 'PERSONAL',
-		name: 'Tâche personnelle',
-		description: 'Créer une tâche uniquement visible par vous',
-	},
-];
+const types = ITEM_TYPES;
 
 const TaskInput = ({onSubmitProject, onSubmitTask, defaultValue}) => {
 	const [value, setValue] = useState(defaultValue);
@@ -160,7 +113,9 @@ const TaskInput = ({onSubmitProject, onSubmitTask, defaultValue}) => {
 	return (
 		<Container ref={ref}>
 			<InputContainer>
-				<Icon onClick={() => setFocusByClick(true)}>{icon}</Icon>
+				<Icon onClick={() => setFocusByClick(true)} active={type}>
+					{icon}
+				</Icon>
 				<Input
 					type="text"
 					onChange={e => setValue(e.target.value)}
